Add resolver for create label mutations; clean up automation lambda

+++ src/api/type-defs/objects/Label.js
@@ -9,7 +9,7 @@ module.exports = `
     type: String!
     category: String!
     conf: Float
-    bbox: [Int!]
+    bbox: [Float!]
     labeledDate: Date!
     validation: Validation!
   }
